simplified capture_token/lookup_permissions and tests
