+++ tests/lookup_permissions.sh
