+++ tests/test_authx.py
@@ -3,6 +3,8 @@ import requests
 import time
 import os
 import json
+from capture_token import helper_get_user_token
+from lookup_permissions import helper_get_permissions
 
 """
 This test suite will cover the manual tests in README.md, ensuring that
@@ -13,50 +15,10 @@ Plus maybe some others that aren't in there
 - modified but live token
 """
 
-OIDC1_URL="http://localhost:8080/auth/realms/mockrealm/protocol/openid-connect"
-OIDC2_URL="http://localhost:8081/auth/realms/mockrealm/protocol/openid-connect"
-
-
-def helper_get_user_token(username, password, oidc_url=OIDC1_URL):
-    client_id = os.getenv("IDP_CLIENT_ID", "mock_login_client")
-    client_secret = os.getenv("IDP_CLIENT_SECRET", "mock_login_secret")
-
-    payload = {'grant_type': 'password',
-               'username': username,
-               'password': password,
-               'redirect_uri': "http://fake_beacon:8000/auth/oidc"}
-
-    response = requests.post(f"{oidc_url}/token", auth=(client_id, client_secret), data=payload)
-    token = response.json()['access_token']
-    return token
-
-
-def helper_get_permissions(token):
-    opa_url = os.getenv("OPAURL", "https://localhost:8181/v1/data/permissions/datasets")
-
-    payload = {
-                'input': {
-                  'headers': {'X-Candig-Local-Oidc': token},
-                  'body': {'method': 'GET', 'path': '/api/phenopackets'}
-                }
-              }
-
-    headers = {
-               'Content-Type': 'application/json',
-               'Accept': 'application/json',
-               'Authorization': 'Bearer my-secret-root-token'
-              }
-
-    # NOTE!  Can't verify https using the certificate because from here 'outside' the
-    # docker network, the hostname of 'opa' is 'localhost'.  You could fix this by
-    # updating /etc/hosts etc.
-    response = requests.post(opa_url, headers=headers, data=json.dumps(payload), verify=False)
-
-    assert response.status_code == 200
-
-    body = response.json()
-    assert "result" in body
-    return body["result"]
+user1 = os.getenv("USER1_ID")
+pass1 = os.getenv("USER1_PASS")
+user2 = os.getenv("USER2_ID")
+pass2 = os.getenv("USER2_PASS")
 
 
 @pytest.fixture(scope="session")
@@ -64,7 +26,7 @@ def user1_token():
     """
     Return the token for user1
     """
-    return helper_get_user_token("user1", "pass1")
+    return helper_get_user_token(user1, pass1)
 
 
 def test_user1_controlled_access(user1_token):
@@ -107,7 +69,7 @@ def user2_token():
     """
     Return the token for user2
     """
-    return helper_get_user_token("user2", "pass2")
+    return helper_get_user_token(user2, pass2)
 
 
 def test_user2_controlled_access(user2_token):
@@ -144,96 +106,8 @@ def test_user2_opt_in_access(user2_token):
     assert "open1" in datasets
     assert "open2" in datasets
 
-@pytest.fixture(scope="session")
-def user3_token():
-    """
-    Return the token for user3
-    """
-    return helper_get_user_token("user3", "pass3", OIDC2_URL)
-
-
-def test_user3_controlled_access(user3_token):
-    """"
-    Make sure user3 has access to controlled4 and controlled6
-    """
-    datasets = helper_get_permissions(user3_token)
-    assert "controlled4" in datasets
-    assert "controlled6" in datasets
-
-
-def test_user3_registered_access(user3_token):
-    """
-    User3, being a trusted researcher, should have acess to registered3
-    """
-    datasets = helper_get_permissions(user3_token)
-    assert "registered3" in datasets
-
-def test_user3_invalid(user3_token):
-    """
-    Make sure invalid token will not have access to datasets other than open datasets
-    """
-    invalid_token = 'A' + user3_token[1:]
-    datasets = helper_get_permissions(invalid_token)
-    assert "controlled4" not in datasets
-    assert "controlled6" not in datasets
-    assert "registered3" not in datasets
-    assert "open1" in datasets
-    assert "open2" in datasets
-
-def test_user3_opt_in_access(user3_token):
-    """
-    Make sure user3 has access to opt in dataset controlled4
-    """
-    datasets = helper_get_permissions(user3_token)
-    assert "controlled4" in datasets
-    assert "open1" in datasets
-    assert "open2" in datasets
-
-
-@pytest.fixture(scope="session")
-def user4_token():
-    """
-    Return the token for user4
-    """
-    return helper_get_user_token("user4", "pass4", OIDC2_URL)
-
-
-def test_user4_controlled_access(user4_token):
-    """"
-    Make sure user4 has access to controlled6 and controlled5
-    """
-    datasets = helper_get_permissions(user4_token)
-    assert "controlled5" in datasets
-
-
-def test_user4_registered_access(user4_token):
-    """
-    User4, not being a trusted researcher, should have acess to registered3
-    """
-    datasets = helper_get_permissions(user4_token)
-    assert "registered3" not in datasets
-
-def test_user4_invalid(user4_token):
-    """
-    Make sure invalid token will not have access to datasets other than open datasets
-    """
-    invalid_token = 'A' + user4_token[1:]
-    datasets = helper_get_permissions(invalid_token)
-    assert "controlled5" not in datasets
-    assert "open1" in datasets
-    assert "open2" in datasets
-
-def test_user4_opt_in_access(user4_token):
-    """
-    Make sure user4 has access to opt in dataset controlled4
-    """
-    datasets = helper_get_permissions(user4_token)
-    assert "controlled5" in datasets
-    assert "open1" in datasets
-    assert "open2" in datasets
-
 
 if __name__ == "__main__":
-    token = helper_get_user_token("user1", "pass1")
+    token = helper_get_user_token(user1, pass1)
     result = helper_get_permissions(token)
     print(result)
